refactor(metrics)!: remove `impl GaugeValue for u32/u64/usize {...}`

+++ examples/src/metrics/process.rs
@@ -14,45 +14,45 @@ use sysinfo::{Pid, ProcessRefreshKind, ProcessesToUpdate, System, get_current_pi
 #[derive(Clone, Register)]
 pub struct ProcessMetrics {
     /// Process ID.
-    pid: ConstGauge<u32>,
+    pid: ConstGauge<i64>,
     /// Total CPU time consumed by the current process in seconds.
     cpu_seconds: LazyGauge<f64>,
     /// CPU usage of the current process in percent.
     cpu_usage_percent: LazyGauge<f32>,
     /// Resident Set Size (RSS) of the current process in bytes.
     #[register(unit(Bytes))]
-    resident_memory: LazyGauge<u64>,
+    resident_memory: LazyGauge<i64>,
     /// Virtual memory size of the current process in bytes.
     #[register(unit(Bytes))]
-    virtual_memory: LazyGauge<u64>,
+    virtual_memory: LazyGauge<i64>,
     /// Process start time in seconds since the Unix epoch.
     #[register(unit(Seconds))]
-    start_time: LazyGauge<u64>,
+    start_time: LazyGauge<i64>,
     /// Process run time in seconds.
     #[register(unit(Seconds))]
-    run_time: LazyGauge<u64>,
+    run_time: LazyGauge<i64>,
     /// Number of open file descriptors for the current process.
-    open_fds: LazyGauge<u32>,
+    open_fds: LazyGauge<i64>,
     /// Limit of open file descriptors for the current process.
-    max_open_fds: LazyGauge<u32>,
+    max_open_fds: LazyGauge<i64>,
     /// Number of threads for the current process.
-    threads: LazyGauge<u32>,
+    threads: LazyGauge<i64>,
 }
 
 impl Default for ProcessMetrics {
     fn default() -> Self {
         let group: LazyGroup<ProcessSample> = LazyGroup::new(|| PROCESS_SAMPLER.sample());
         Self {
-            pid: ConstGauge::new(PROCESS_SAMPLER.pid.as_u32()),
+            pid: ConstGauge::new(PROCESS_SAMPLER.pid.as_u32() as i64),
             cpu_seconds: group.gauge(|s| s.cpu_seconds_total),
             cpu_usage_percent: group.gauge(|s| s.cpu_usage_percent),
-            resident_memory: group.gauge(|s| s.resident_memory_bytes),
-            virtual_memory: group.gauge(|s| s.virtual_memory_bytes),
-            start_time: group.gauge(|s| s.start_time_seconds),
-            run_time: group.gauge(|s| s.run_time_seconds),
-            open_fds: group.gauge(|s| s.open_fds),
-            max_open_fds: group.gauge(|s| s.max_open_fds),
-            threads: group.gauge(|s| s.thread_count),
+            resident_memory: group.gauge(|s| s.resident_memory_bytes as i64),
+            virtual_memory: group.gauge(|s| s.virtual_memory_bytes as i64),
+            start_time: group.gauge(|s| s.start_time_seconds as i64),
+            run_time: group.gauge(|s| s.run_time_seconds as i64),
+            open_fds: group.gauge(|s| s.open_fds as i64),
+            max_open_fds: group.gauge(|s| s.max_open_fds as i64),
+            threads: group.gauge(|s| s.thread_count as i64),
         }
     }
 }
